all: adapt all cover and sig to 64bit
Taken some arm64 devices for example:
kaslr_offset is diff at bits 12-40, and kernel modules are loaded at 2GB space,
so we have `ffffffd342e10000 T _stext` where uppper 32bit is ffffffd3. However,
if we check modules range, the 1st module is loaded at 0xffffffd2eeb2a000,
while the last module is loaded at 0xffffffd2f42c4000.
We can see the upper 32bits are diff for core kernel and modules.

If we use current 32bits for covered PC, we will get wrong module address
recovered.

So we need to move to 64bit cover and signal:
- change cover/sig to 64bit to fit for syz-executor change
- remove kernel upper base logic as kernel upper base is not a constant when
  kaslr enabled for core kernel and modules.
- remove unused pcBase

+++ pkg/cover/backend/backend.go
@@ -14,7 +14,7 @@ type Impl struct {
 	Symbols         []*Symbol
 	Frames          []Frame
 	Symbolize       func(pcs map[*Module][]uint64) ([]Frame, error)
-	RestorePC       func(pc uint32) uint64
+	RestorePC       func(pc uint64) uint64
 	CallbackPoints  []uint64
 	PreciseCoverage bool
 }
